requirements.txt
PyYAML==3.13
sqlalchemy==1.2.12
flask==1.0.2
flask-restful==0.3.6
flask_sqlalchemy==2.3.2
flask_script==2.0.6
flask_migrate==2.3.0
psycopg2-binary==2.7.5
requests==2.20.0
marshmallow==2.16.1
flask-cors==3.0.6
pandas==0.23.4
pydash==4.7.3
apispec==0.39.0
flask_apispec==0.7.0